D3D12 backend scaffolding and Vortice upgrade progress
Agent-Logs-Url: https://github.com/winnerspiros/veldrid/sessions/3eb551ab-579b-420a-89e2-6a532bed6865

+++ src/Veldrid/D3D11/D3D11Framebuffer.cs
@@ -45,7 +45,7 @@ namespace Veldrid.D3D11
                     if (d3dDepthTarget.SampleCount == TextureSampleCount.Count1)
                     {
                         dsvDesc.ViewDimension = DepthStencilViewDimension.Texture2D;
-                        dsvDesc.Texture2D.MipSlice = (int)description.DepthTarget.Value.MipLevel;
+                        dsvDesc.Texture2D.MipSlice = description.DepthTarget.Value.MipLevel;
                     }
                     else
                         dsvDesc.ViewDimension = DepthStencilViewDimension.Texture2DMultisampled;
@@ -55,14 +55,14 @@ namespace Veldrid.D3D11
                     if (d3dDepthTarget.SampleCount == TextureSampleCount.Count1)
                     {
                         dsvDesc.ViewDimension = DepthStencilViewDimension.Texture2DArray;
-                        dsvDesc.Texture2DArray.FirstArraySlice = (int)description.DepthTarget.Value.ArrayLayer;
+                        dsvDesc.Texture2DArray.FirstArraySlice = description.DepthTarget.Value.ArrayLayer;
                         dsvDesc.Texture2DArray.ArraySize = 1;
-                        dsvDesc.Texture2DArray.MipSlice = (int)description.DepthTarget.Value.MipLevel;
+                        dsvDesc.Texture2DArray.MipSlice = description.DepthTarget.Value.MipLevel;
                     }
                     else
                     {
                         dsvDesc.ViewDimension = DepthStencilViewDimension.Texture2DMultisampledArray;
-                        dsvDesc.Texture2DMSArray.FirstArraySlice = (int)description.DepthTarget.Value.ArrayLayer;
+                        dsvDesc.Texture2DMSArray.FirstArraySlice = description.DepthTarget.Value.ArrayLayer;
                         dsvDesc.Texture2DMSArray.ArraySize = 1;
                     }
                 }
@@ -90,8 +90,8 @@ namespace Veldrid.D3D11
                             rtvDesc.Texture2DArray = new Texture2DArrayRenderTargetView
                             {
                                 ArraySize = 1,
-                                FirstArraySlice = (int)description.ColorTargets[i].ArrayLayer,
-                                MipSlice = (int)description.ColorTargets[i].MipLevel
+                                FirstArraySlice = description.ColorTargets[i].ArrayLayer,
+                                MipSlice = description.ColorTargets[i].MipLevel
                             };
                         }
                         else
@@ -100,7 +100,7 @@ namespace Veldrid.D3D11
                             rtvDesc.Texture2DMSArray = new Texture2DMultisampledArrayRenderTargetView
                             {
                                 ArraySize = 1,
-                                FirstArraySlice = (int)description.ColorTargets[i].ArrayLayer
+                                FirstArraySlice = description.ColorTargets[i].ArrayLayer
                             };
                         }
                     }
@@ -109,7 +109,7 @@ namespace Veldrid.D3D11
                         if (d3dColorTarget.SampleCount == TextureSampleCount.Count1)
                         {
                             rtvDesc.ViewDimension = RenderTargetViewDimension.Texture2D;
-                            rtvDesc.Texture2D.MipSlice = (int)description.ColorTargets[i].MipLevel;
+                            rtvDesc.Texture2D.MipSlice = description.ColorTargets[i].MipLevel;
                         }
                         else
                             rtvDesc.ViewDimension = RenderTargetViewDimension.Texture2DMultisampled;
