+++ src/Veldrid/D3D11/D3D11Sampler.cs
@@ -35,7 +35,7 @@ namespace Veldrid.D3D11
                 Filter = D3D11Formats.ToD3D11Filter(description.Filter, description.ComparisonKind.HasValue),
                 MinLOD = description.MinimumLod,
                 MaxLOD = description.MaximumLod,
-                MaxAnisotropy = (int)description.MaximumAnisotropy,
+                MaxAnisotropy = description.MaximumAnisotropy,
                 ComparisonFunc = comparision,
                 MipLODBias = description.LodBias,
                 BorderColor = toRawColor4(description.BorderColor)
